Moved LocationArgumentResolver to ControllerArgumentResolved namespace
