+++ src/bundle/Core/Resources/config/services.yml
@@ -195,11 +195,11 @@ services:
         tags:
             - { name: request.param_converter, priority: '%ibexa.param_converter.location.priority%', converter: ez_location_converter }
 
-    Ibexa\Bundle\Core\Converter\LocationArgumentResolver:
+    Ibexa\Bundle\Core\ControllerArgumentResolver\LocationArgumentResolver:
         autowire: true
         autoconfigure: true
         tags:
-            - { name: controller.argument_value_resolver, priority: '%ibexa.param_converter.location.priority%' }
+            - { name: controller.argument_value_resolver, priority: 50 }
 
     Ibexa\Bundle\Core\EventListener\ExceptionListener:
         class: Ibexa\Bundle\Core\EventListener\ExceptionListener
